split off an EncoderInterface and NormalizerInterface from SerializerInterface
