+++ src/Symfony/Component/Serializer/EncoderInterface.php
@@ -1,0 +1,63 @@
+<?php
+
+namespace Symfony\Component\Serializer;
+
+use Symfony\Component\Serializer\Encoder\EncoderInterface;
+
+/*
+ * This file is part of the Symfony framework.
+ *
+ * (c) Fabien Potencier <fabien@symfony.com>
+ *
+ * This source file is subject to the MIT license that is bundled
+ * with this source code in the file LICENSE.
+ */
+
+/**
+ * Defines the interface of the Encoder
+ *
+ * @author Jordi Boggiano <j.boggiano@seld.be>
+ */
+interface EncoderInterface
+{
+    /**
+     * Encodes data into the given format
+     *
+     * @param mixed $data data to encode
+     * @param string $format format name
+     * @return array|scalar
+     */
+    function encode($data, $format);
+
+    /**
+     * Decodes a string from the given format back into PHP data
+     *
+     * @param string $data data to decode
+     * @param string $format format name
+     * @return mixed
+     */
+    function decode($data, $format);
+
+    /**
+     * Checks whether the serializer can encode to given format
+     *
+     * @param string $format format name
+     * @return Boolean
+     */
+    function supportsEncoding($format);
+
+    /**
+     * Checks whether the serializer can decode from given format
+     *
+     * @param string $format format name
+     * @return Boolean
+     */
+    function supportsDecoding($format);
+
+    /**
+     * Get the encoder for the given format
+     *
+     * @return EncoderInterface
+     */
+    function getEncoder($format);
+}
