+++ src/Symfony/Component/Serializer/NormalizerInterface.php
@@ -1,0 +1,41 @@
+<?php
+
+namespace Symfony\Component\Serializer;
+
+use Symfony\Component\Serializer\Encoder\EncoderInterface;
+
+/*
+ * This file is part of the Symfony framework.
+ *
+ * (c) Fabien Potencier <fabien@symfony.com>
+ *
+ * This source file is subject to the MIT license that is bundled
+ * with this source code in the file LICENSE.
+ */
+
+/**
+ * Defines the interface of the Normalizer
+ *
+ * @author Jordi Boggiano <j.boggiano@seld.be>
+ */
+interface NormalizerInterface
+{
+    /**
+     * Normalizes any data into a set of arrays/scalars
+     *
+     * @param mixed $data data to normalize
+     * @param string $format format name, present to give the option to normalizers to act differently based on formats
+     * @return array|scalar
+     */
+    function normalize($data, $format = null);
+
+    /**
+     * Denormalizes data into the given type.
+     *
+     * @param mixed $data
+     * @param string $type
+     * @param string $format
+     * @return mixed
+     */
+    function denormalize($data, $type, $format = null);
+}
