+++ src/Symfony/Component/Serializer/SerializerInterface.php
@@ -39,43 +39,6 @@ interface SerializerInterface
     function deserialize($data, $type, $format);
 
     /**
-     * Normalizes any data into a set of arrays/scalars
-     *
-     * @param mixed $data data to normalize
-     * @param string $format format name, present to give the option to normalizers to act differently based on formats
-     * @return array|scalar
-     */
-    function normalize($data, $format = null);
-
-    /**
-     * Denormalizes data into the given type.
-     *
-     * @param mixed $data
-     * @param string $type
-     * @param string $format
-     * @return mixed
-     */
-    function denormalize($data, $type, $format = null);
-
-    /**
-     * Encodes data into the given format
-     *
-     * @param mixed $data data to encode
-     * @param string $format format name
-     * @return array|scalar
-     */
-    function encode($data, $format);
-
-    /**
-     * Decodes a string from the given format back into PHP data
-     *
-     * @param string $data data to decode
-     * @param string $format format name
-     * @return mixed
-     */
-    function decode($data, $format);
-
-    /**
      * Checks whether the serializer can serialize to given format
      *
      * @param string $format format name
@@ -90,27 +53,4 @@ interface SerializerInterface
      * @return Boolean
      */
     function supportsDeserialization($format);
-
-    /**
-     * Checks whether the serializer can encode to given format
-     *
-     * @param string $format format name
-     * @return Boolean
-     */
-    function supportsEncoding($format);
-
-    /**
-     * Checks whether the serializer can decode from given format
-     *
-     * @param string $format format name
-     * @return Boolean
-     */
-    function supportsDecoding($format);
-
-    /**
-     * Get the encoder for the given format
-     *
-     * @return EncoderInterface
-     */
-    function getEncoder($format);
 }
